requirements.txt
crawl4ai
openai
python-dotenv
pandas
playwright
groq